MacOS flaky tests: remove isDarwin()
* log when tests are being skipped
* set CI env var for MacOS github runner
* fix warning we cant assign null: `let secondContext: osn.IVideo = null;`
* do not invoke `AddVideoContext` on CI builds in enhanced broadcasting tests

+++ tests/osn-tests/src/test_osn_enhanced_broadcasting_simple_streaming.ts
@@ -12,11 +12,11 @@ import path = require('path');
 
 const testName = 'osn-enhanced-broadcasting-simple-streaming';
 
-describe(testName, () => {
+describe(testName, function() {
     let obs: OBSHandler;
     let hasTestFailed: boolean = false;
     const mediaPath = path.join(path.normalize(__dirname), '..', 'media');
-    let secondContext: osn.IVideo = null;
+    let secondContext: osn.IVideo;
 
     // Initialize OBS process
     before(async() => {
@@ -29,21 +29,23 @@ describe(testName, () => {
         // Reserving user from pool
         await obs.reserveUser();
 
-        secondContext = osn.VideoFactory.create();
-        const secondVideoInfo: osn.IVideoInfo = {
-            fpsNum: 60,
-            fpsDen: 2,
-            baseWidth: 720,
-            baseHeight: 1280,
-            outputWidth: 720,
-            outputHeight: 1280,
-            outputFormat: osn.EVideoFormat.NV12,
-            colorspace: osn.EColorSpace.CS709,
-            range: osn.ERangeType.Full,
-            scaleType: osn.EScaleType.Lanczos,
-            fpsType: osn.EFPSType.Fractional
-        };
-        secondContext.video = secondVideoInfo;
+        if (!obs.isCI()) {
+            secondContext = osn.VideoFactory.create();
+            const secondVideoInfo: osn.IVideoInfo = {
+                fpsNum: 60,
+                fpsDen: 2,
+                baseWidth: 720,
+                baseHeight: 1280,
+                outputWidth: 720,
+                outputHeight: 1280,
+                outputFormat: osn.EVideoFormat.NV12,
+                colorspace: osn.EColorSpace.CS709,
+                range: osn.ERangeType.Full,
+                scaleType: osn.EScaleType.Lanczos,
+                fpsType: osn.EFPSType.Fractional
+            };
+            secondContext.video = secondVideoInfo;
+        }
     });
 
     // Shutdown OBS process
@@ -51,7 +53,9 @@ describe(testName, () => {
         // Releasing user got from pool
         await obs.releaseUser();
 
-        secondContext.destroy();
+        if (secondContext) {
+            secondContext.destroy();
+        }
         obs.shutdown();
 
         if (hasTestFailed === true) {
@@ -70,11 +74,8 @@ describe(testName, () => {
     });
 
     it('Enhanced Broadcasting Simple Streaming Single Canvas', async function() {
-        if (obs.isDarwin()) {
-            this.skip();
-        }
-
         if (obs.isCI()) {
+            logInfo(testName, 'Running in CI environment, skipping test that requires GPU');
             // Skipping this test because CI server doesn't have GPU, but you can run it locally
             this.skip();
         }
@@ -166,11 +167,8 @@ describe(testName, () => {
     });
 
     it('Enhanced Broadcasting Simple Streaming Dual Canvas', async function() {
-        if (obs.isDarwin()) {
-            this.skip();
-        }
-
         if (obs.isCI()) {
+            logInfo(testName, 'Running in CI environment, skipping test that requires GPU');
             // Skipping this test because CI server doesn't have GPU, but you can run it locally
             this.skip();
         }
